+++ src/slaynash/sgengine/gui/GUIProgressBar.java
@@ -2,8 +2,8 @@ package slaynash.sgengine.gui;
 
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL30;
-import org.lwjgl.util.vector.Vector2f;
 
+import slaynash.sgengine.maths.Vector2i;
 import slaynash.sgengine.models.Renderable2dModel;
 import slaynash.sgengine.models.utils.VaoManager;
 import slaynash.sgengine.shaders.ShaderManager;
@@ -55,7 +55,7 @@ public class GUIProgressBar extends GUIElement{
 	public void render() {
 		ShaderManager.shader_loadTranslation(getTopLeft());
 		model.render();
-		ShaderManager.shader_loadTranslation(new Vector2f());
+		ShaderManager.shader_loadTranslation(new Vector2i());
 		/*
 		GL11.glBegin(GL11.GL_QUADS);
 			GL11.glTexCoord2f(0      , 0);
